Introduced --localOnly option so namespace doesn't need to be specified
Previously, to make the controller only operate in a single namespace,
you had to specify the namespace using --namespace. This required the
per-namespace.yaml to include the name of the namespace, which wasn't
ideal. Now, you can simply run the controller with --localOnly and it
will operate on StatefulSets in the namespace the controller is running
in (the namespace is read from /var/run/secrets/kubernetes.io/serviceaccount/namespace)