Ensure only one drain pod is created when PodManagementPolicy is OrderedReady
If PodManagementPolicy is OrderedReady, the controller should only
create one drain pod, wait for it to finish, and then create the second
drain pod if the StatefulSet was scaled down by 2 or more.

If the PodManagementPolicy is Parallel, the controller should create
all the required drain pods (e.g. two when StatefulSet is scaled down
by 2).